refactor state management to use context provider component
