+++ src/components/App/App.js
@@ -1,15 +1,15 @@
-import React from 'react';
-
-import ToastPlayground from '../ToastPlayground';
-import Footer from '../Footer';
+import React from 'react'
+import ToastProvider from '../ToastProvider'
+import ToastPlayground from '../ToastPlayground'
+import Footer from '../Footer'
 
 function App() {
   return (
-    <>
+    <ToastProvider>
       <ToastPlayground />
       <Footer />
-    </>
-  );
+    </ToastProvider>
+  )
 }
 
-export default App;
+export default App
